Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -16,7 +16,7 @@ int main(){
      else if (z == '/') {
         // Cast x to float or use float for the division result
         if (y != 0) {
-            printf("%.2f", (float)x / y);  // Using %.2f for a rounded float output
+            printf("%d", x / y);  // Using %.2f for a rounded float output
         } else {
             printf("error");
         }
